+++ modules/Geometry/src/Device.cpp
@@ -25,7 +25,7 @@ std::shared_ptr<PointCloud> Device::pointsInFov(const PointCloud &pcd) const
 
         float horizontal_d_angle = std::atan2(transformNode_->getLocalTransform().get3DDirectionVector().y(), transformNode_->getLocalTransform().get3DDirectionVector().x());
         float vertical_d_angle = std::atan2(transformNode_->getLocalTransform().get3DDirectionVector().y(), transformNode_->getLocalTransform().get3DDirectionVector().z());
-        float epsilon = 1e-7;
+        float epsilon = static_cast<float>(1e-7);
         if ((std::abs(horizontal_p_angle - horizontal_d_angle) - (horizontal_fov_rad_ / 2.0f)) < epsilon &&
             (std::abs(vertical_p_angle - vertical_d_angle) - (vertical_fov_rad_ / 2.0f)) < epsilon)
         {
